backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
flask-cors==4.0.1
boto3==1.28.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.86.0
python-dotenv==1.0.0
